MEM.EXE: Query the heap for the top of memory
This program formerly used MEM/MEM.LIB to find the top of memory.
MEM/MEM.LIB is identical to the Microsoft C standard library
(src/TOOLS/BLD/LIB/SLIBCE.LIB) except that the startup module is
patched to provide the top of available memory.

This change adds the function TopOfMemory to chase pointers
through the heap and find the top of memory. This removes the
dependency on MEM/MEM.LIB so that this does not have to be
re-created for the Watcom compiler.